Moved navigation-behavior and fixed key issue in copy-coordinates menu
