+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/navigation-behavior/index.tsx
