+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/navigation-behavior/navigation-behavior.tsx
